+++ spec/features/swagger_spec.rb
@@ -19,14 +19,24 @@ describe 'Swagger' do
     context "#headers" do
       before do
         GrapeSwaggerRails.options.headers['X-Test-Header'] = 'Test Value'
+        GrapeSwaggerRails.options.headers['X-Another-Header'] = 'Another Value'
         visit '/swagger'
       end
       it 'adds headers' do
         find('#endpointListTogger_headers', visible: true).click
-        find('a[href="#!/headers/GET_api_headers_format"]', visible: true).click
+        first('a[href="#!/headers/GET_api_headers_format"]', visible: true).click
         click_button 'Try it out!'
         expect(page).to have_css "span.attribute", text: 'X-Test-Header'
         expect(page).to have_css "span.string", text: 'Test Value'
+      end
+      it 'supports multiple headers' do
+        find('#endpointListTogger_headers', visible: true).click
+        first('a[href="#!/headers/GET_api_headers_format"]', visible: true).click
+        click_button 'Try it out!'
+        expect(page).to have_css "span.attribute", text: 'X-Test-Header'
+        expect(page).to have_css "span.string", text: 'Test Value'
+        expect(page).to have_css "span.attribute", text: 'X-Another-Header'
+        expect(page).to have_css "span.string", text: 'Another Value'
       end
     end
     context "#api_auth:basic" do