+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Next
 
+* [#6](https://github.com/TinkerDev/grape-swagger-rails/pull/6): Fix: support multiple predefined headers - [@Tyr0](https://github.com/tyr0).
 * Your contribution here.
 
 * Added a new :custom_css_file option to have default index.html.erb render a stylesheet link for custom theming - [@alexagranov](https://github.com/alexagranov).